tidy constants
