+++ packages/insomnia/src/insomnia-data/src/models/request.ts
@@ -15,10 +15,10 @@
 
 import { OperationTypeNode } from 'graphql';
 
+import type { OAuth1SignatureMethod } from '~/common/constants';
 import { METHOD_GET } from '~/common/constants';
 import { replaceIdsInFields } from '~/models/helpers/replace-ids-in-fields';
 import type { BaseModel } from '~/models/types';
-import type { OAuth1SignatureMethod } from '~/network/o-auth-1/constants';
 import { getOperationType } from '~/utils/graph-ql';
 
 export const name = 'Request';
