+++ packages/insomnia/src/main/network/o-auth-1/get-token.ts
@@ -11,7 +11,7 @@ import {
   SIGNATURE_METHOD_HMAC_SHA256,
   SIGNATURE_METHOD_PLAINTEXT,
   SIGNATURE_METHOD_RSA_SHA1,
-} from '../../../network/o-auth-1/constants';
+} from './constants';
 
 function hashFunction(signatureMethod: OAuth1SignatureMethod) {
   if (signatureMethod === SIGNATURE_METHOD_HMAC_SHA1) {
