+++ packages/insomnia/src/ui/components/modals/oauth-authorization-status-modal.tsx
@@ -1,8 +1,8 @@
 import React, { type FC, useEffect, useRef, useState } from 'react';
 
+import type { OAuth2AuthorizationStatusType } from '~/common/constants';
 import { useDefaultBrowserRedirectActionFetcher } from '~/routes/auth.default-browser-redirect';
 
-import type { OAuth2AuthorizationStatusType } from '../../../network/o-auth-2/constants';
 import { invariant } from '../../../utils/invariant';
 import uiEventBus, { OAUTH2_AUTHORIZATION_STATUS_CHANGE } from '../../event-bus';
 import { Modal, type ModalHandle } from '../base/modal';
